+++ crates/ryo3-reqwest/src/client.rs
@@ -1,9 +1,11 @@
 use crate::RyResponse;
 use crate::errors::map_reqwest_err;
+use crate::tls_version::TlsVersion;
 use crate::user_agent::parse_user_agent;
 use bytes::Bytes;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
+use pyo3::pybacked::PyBackedStr;
 use pyo3::types::{PyDict, PyTuple};
 use pyo3::{IntoPyObjectExt, intern};
 use reqwest::header::HeaderMap;
@@ -27,15 +29,70 @@ pub struct ClientConfig {
     headers: Option<PyHeaders>,
     cookies: bool,
     user_agent: Option<ryo3_http::HttpHeaderValue>,
+    hickory_dns: bool,
+    redirect: Option<usize>,
+    referer: bool,
+    // -- http preferences --
+    http1_only: bool,
+    https_only: bool,
+    // -- http1 --
+    http1_title_case_headers: bool,
+    http1_allow_obsolete_multiline_headers_in_responses: bool,
+    http1_allow_spaces_after_header_name_in_responses: bool,
+    http1_ignore_invalid_headers_in_responses: bool,
+    // -- http2 --
+    http2_prior_knowledge: bool,
+    http2_initial_stream_window_size: Option<u32>,
+    http2_initial_connection_window_size: Option<u32>,
+    http2_adaptive_window: bool,
+    http2_max_frame_size: Option<u32>,
+    http2_max_header_list_size: Option<u32>,
+    http2_keep_alive_interval: Option<PyDuration>,
+    http2_keep_alive_timeout: Option<PyDuration>,
+    http2_keep_alive_while_idle: bool,
+    // -- timeout(s) --
     timeout: Option<PyDuration>,
     read_timeout: Option<PyDuration>,
     connect_timeout: Option<PyDuration>,
+    // -- compression --
     gzip: bool,
     brotli: bool,
     deflate: bool,
     zstd: bool,
-    hickory_dns: bool,
-    http1_only: bool,
+    // -- pool --
+    pool_max_idle_per_host: usize,
+    pool_idle_timeout: Option<PyDuration>,
+    // -- tcp --
+    tcp_keepalive: Option<PyDuration>, // default: 15 seconds
+    tcp_keepalive_interval: Option<PyDuration>, // default: 15 seconds
+    tcp_keepalive_retries: Option<u32>, // default: 3
+    tcp_nodelay: bool,                 // default: true
+    // -- tls --
+    tls_max_version: Option<TlsVersion>,
+    tls_min_version: Option<TlsVersion>,
+    tls_info: bool, // default: false
+    tls_sni: bool,  // default: true
+    // -- danger zone --
+    danger_accept_invalid_certs: bool,
+    danger_accept_invalid_hostnames: bool,
+    // == CLIENT BUILDER OPTIONS TODO ==
+    // add_crl
+    // add_crls
+    // add_root_certificate
+    // connector_layer
+    // cookie_provider
+    // cookie_store
+    // dns_resolver
+    // dns_resolver2
+    // http09_responses
+    // identity
+    // interface
+    // local_address
+    // proxy
+    // referer
+    // resolve
+    // resolve_to_addrs
+    // retry
 }
 
 struct RequestKwargs<'py> {
@@ -48,6 +105,8 @@ struct RequestKwargs<'py> {
     multipart: Option<&'py Bound<'py, PyAny>>,
     form: Option<&'py Bound<'py, PyAny>>,
     timeout: Option<&'py PyDuration>,
+    basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+    bearer_auth: Option<PyBackedStr>,
     version: Option<HttpVersion>,
 }
 
@@ -64,6 +123,12 @@ impl RyHttpClient {
     fn build_request<'py>(&'py self, options: RequestKwargs<'py>) -> PyResult<RequestBuilder> {
         let url = extract_url(options.url)?;
         let mut req = self.client.request(options.method, url);
+        if let Some((username, password)) = options.basic_auth {
+            req = req.basic_auth(username, password);
+        }
+        if let Some(token) = options.bearer_auth {
+            req = req.bearer_auth(token);
+        }
         if let Some(ref version) = options.version {
             req = req.version(version.0);
         }
@@ -118,7 +183,11 @@ impl RyHttpClient {
 
 #[pymethods]
 impl RyHttpClient {
-    #[expect(clippy::too_many_arguments)]
+    #[expect(
+        clippy::fn_params_excessive_bools,
+        clippy::similar_names,
+        clippy::too_many_arguments
+    )]
     #[new]
     #[pyo3(
         signature = (
@@ -129,12 +198,47 @@ impl RyHttpClient {
             timeout = None,
             read_timeout = None,
             connect_timeout = None,
+            redirect = Some(10),
+            referer = true,
             gzip = true,
             brotli = true,
             deflate = true,
             zstd = true,
             hickory_dns = true,
+
             http1_only = false,
+            https_only = false,
+
+            http1_title_case_headers = false,
+            http1_allow_obsolete_multiline_headers_in_responses = false,
+            http1_allow_spaces_after_header_name_in_responses = false,
+            http1_ignore_invalid_headers_in_responses = false,
+
+            http2_prior_knowledge = false,
+            http2_initial_stream_window_size = None,
+            http2_initial_connection_window_size = None,
+            http2_adaptive_window = false,
+            http2_max_frame_size = None,
+            http2_max_header_list_size = None,
+            http2_keep_alive_interval = None,
+            http2_keep_alive_timeout = None,
+            http2_keep_alive_while_idle = false,
+
+            pool_idle_timeout = Some(PyDuration::from(std::time::Duration::from_secs(90))),
+            pool_max_idle_per_host = usize::MAX,
+
+            tcp_keepalive = Some(PyDuration::from(std::time::Duration::from_secs(15))),
+            tcp_keepalive_interval = Some(PyDuration::from(std::time::Duration::from_secs(15))),
+            tcp_keepalive_retries = Some(3),
+            tcp_nodelay = true,
+
+            tls_min_version = None,
+            tls_max_version = None,
+            tls_info = false,
+            tls_sni = true,
+
+            danger_accept_invalid_certs = false,
+            danger_accept_invalid_hostnames = false,
         )
     )]
     fn py_new(
@@ -144,12 +248,52 @@ impl RyHttpClient {
         timeout: Option<PyDuration>,
         read_timeout: Option<PyDuration>,
         connect_timeout: Option<PyDuration>,
+        redirect: Option<usize>,
+        referer: bool,
         gzip: Option<bool>,
         brotli: Option<bool>,
         deflate: Option<bool>,
         zstd: Option<bool>,
         hickory_dns: Option<bool>,
         http1_only: Option<bool>,
+        https_only: Option<bool>,
+
+        // -- http1 --
+        http1_title_case_headers: bool,
+        http1_allow_obsolete_multiline_headers_in_responses: bool,
+        http1_allow_spaces_after_header_name_in_responses: bool,
+        http1_ignore_invalid_headers_in_responses: bool,
+
+        // -- http2 --
+        http2_prior_knowledge: bool,
+        http2_initial_stream_window_size: Option<u32>,
+        http2_initial_connection_window_size: Option<u32>,
+        http2_adaptive_window: bool,
+        http2_max_frame_size: Option<u32>,
+        http2_max_header_list_size: Option<u32>,
+        http2_keep_alive_interval: Option<PyDuration>,
+        http2_keep_alive_timeout: Option<PyDuration>,
+        http2_keep_alive_while_idle: bool,
+
+        // -- pool --
+        pool_idle_timeout: Option<PyDuration>,
+        pool_max_idle_per_host: usize,
+
+        // -- tcp --
+        tcp_keepalive: Option<PyDuration>,
+        tcp_keepalive_interval: Option<PyDuration>,
+        tcp_keepalive_retries: Option<u32>,
+        tcp_nodelay: bool,
+
+        // -- tls --
+        tls_min_version: Option<TlsVersion>,
+        tls_max_version: Option<TlsVersion>,
+        tls_info: bool,
+        tls_sni: bool,
+
+        // -- danger --
+        danger_accept_invalid_certs: bool,
+        danger_accept_invalid_hostnames: bool,
     ) -> PyResult<Self> {
         let user_agent = parse_user_agent(user_agent)?;
         let headers = headers.map(PyHeaders::try_from).transpose()?;
@@ -160,12 +304,46 @@ impl RyHttpClient {
             timeout,
             read_timeout,
             connect_timeout,
+            redirect,
+            referer,
             gzip: gzip.unwrap_or(true),
             brotli: brotli.unwrap_or(true),
             deflate: deflate.unwrap_or(true),
             zstd: zstd.unwrap_or(true),
             hickory_dns: hickory_dns.unwrap_or(true),
             http1_only: http1_only.unwrap_or(false),
+            https_only: https_only.unwrap_or(false),
+            // -- http1 --
+            http1_title_case_headers,
+            http1_allow_obsolete_multiline_headers_in_responses,
+            http1_allow_spaces_after_header_name_in_responses,
+            http1_ignore_invalid_headers_in_responses,
+            // -- http2 --
+            http2_prior_knowledge,
+            http2_initial_stream_window_size,
+            http2_initial_connection_window_size,
+            http2_adaptive_window,
+            http2_max_frame_size,
+            http2_max_header_list_size,
+            http2_keep_alive_interval,
+            http2_keep_alive_timeout,
+            http2_keep_alive_while_idle,
+            // --- pool ---
+            pool_idle_timeout,
+            pool_max_idle_per_host,
+            // --- tcp ---
+            tcp_keepalive,
+            tcp_keepalive_interval,
+            tcp_keepalive_retries,
+            tcp_nodelay,
+            // --- TLS ---
+            tls_min_version,
+            tls_max_version,
+            tls_info,
+            tls_sni,
+            // -- danger --
+            danger_accept_invalid_certs,
+            danger_accept_invalid_hostnames,
         };
         let client_builder = client_cfg.client_builder();
         let client = client_builder
@@ -211,6 +389,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -226,6 +406,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -238,6 +420,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -260,6 +444,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -275,6 +461,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -287,6 +475,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -309,6 +499,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -324,6 +516,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -336,6 +530,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -358,6 +554,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -373,6 +571,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -385,6 +585,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -407,6 +609,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -422,6 +626,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -434,6 +640,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -456,6 +664,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -471,6 +681,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -483,6 +695,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -505,6 +719,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -520,6 +736,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let opts = RequestKwargs {
@@ -532,6 +750,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
         let req = self.build_request(opts)?;
@@ -555,6 +775,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -571,6 +793,8 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         let method = method.map_or_else(|| Method::GET, |m| m.0);
@@ -584,6 +808,8 @@ impl RyHttpClient {
             multipart,
             form,
             timeout,
+            basic_auth,
+            bearer_auth,
             version,
         };
 
@@ -608,6 +834,8 @@ impl RyHttpClient {
             form = None,
             multipart = None,
             timeout = None,
+            basic_auth = None,
+            bearer_auth = None,
             version = None,
         )
     )]
@@ -624,10 +852,24 @@ impl RyHttpClient {
         form: Option<&Bound<'py, PyAny>>,
         multipart: Option<&Bound<'py, PyAny>>,
         timeout: Option<&PyDuration>,
+        basic_auth: Option<(PyBackedStr, Option<PyBackedStr>)>,
+        bearer_auth: Option<PyBackedStr>,
         version: Option<HttpVersion>,
     ) -> PyResult<Bound<'py, PyAny>> {
         self.fetch(
-            py, url, method, body, headers, query, json, form, multipart, timeout, version,
+            py,
+            url,
+            method,
+            body,
+            headers,
+            query,
+            json,
+            form,
+            multipart,
+            timeout,
+            basic_auth,
+            bearer_auth,
+            version,
         )
     }
 }
@@ -640,16 +882,103 @@ impl<'py> IntoPyObject<'py> for &ClientConfig {
     fn into_pyobject(self, py: Python<'py>) -> Result<Self::Output, Self::Error> {
         let dict = PyDict::new(py);
         dict.set_item(intern!(py, "headers"), self.headers.clone())?;
+        dict.set_item(intern!(py, "cookies"), self.cookies)?;
         dict.set_item(intern!(py, "user_agent"), self.user_agent.clone())?;
         dict.set_item(intern!(py, "timeout"), self.timeout)?;
         dict.set_item(intern!(py, "read_timeout"), self.read_timeout)?;
         dict.set_item(intern!(py, "connect_timeout"), self.connect_timeout)?;
+        dict.set_item(intern!(py, "redirect"), self.redirect)?;
+        dict.set_item(intern!(py, "referer"), self.referer)?;
         dict.set_item(intern!(py, "gzip"), self.gzip)?;
         dict.set_item(intern!(py, "brotli"), self.brotli)?;
         dict.set_item(intern!(py, "deflate"), self.deflate)?;
         dict.set_item(intern!(py, "zstd"), self.zstd)?;
-        dict.set_item(intern!(py, "http1_only"), self.http1_only)?;
         dict.set_item(intern!(py, "hickory_dns"), self.hickory_dns)?;
+        dict.set_item(intern!(py, "http1_only"), self.http1_only)?;
+        // -- http1 --
+        dict.set_item(
+            intern!(py, "http1_title_case_headers"),
+            self.http1_title_case_headers,
+        )?;
+        dict.set_item(
+            intern!(py, "http1_allow_obsolete_multiline_headers_in_responses"),
+            self.http1_allow_obsolete_multiline_headers_in_responses,
+        )?;
+        dict.set_item(
+            intern!(py, "http1_allow_spaces_after_header_name_in_responses"),
+            self.http1_allow_spaces_after_header_name_in_responses,
+        )?;
+        dict.set_item(
+            intern!(py, "http1_ignore_invalid_headers_in_responses"),
+            self.http1_ignore_invalid_headers_in_responses,
+        )?;
+        // -- http2 --
+        dict.set_item(
+            intern!(py, "http2_prior_knowledge"),
+            self.http2_prior_knowledge,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_initial_stream_window_size"),
+            self.http2_initial_stream_window_size,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_initial_connection_window_size"),
+            self.http2_initial_connection_window_size,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_adaptive_window"),
+            self.http2_adaptive_window,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_max_frame_size"),
+            self.http2_max_frame_size,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_max_header_list_size"),
+            self.http2_max_header_list_size,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_keep_alive_interval"),
+            self.http2_keep_alive_interval,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_keep_alive_timeout"),
+            self.http2_keep_alive_timeout,
+        )?;
+        dict.set_item(
+            intern!(py, "http2_keep_alive_while_idle"),
+            self.http2_keep_alive_while_idle,
+        )?;
+        // -- pool --
+        dict.set_item(intern!(py, "pool_idle_timeout"), self.pool_idle_timeout)?;
+        dict.set_item(
+            intern!(py, "pool_max_idle_per_host"),
+            self.pool_max_idle_per_host,
+        )?;
+        // -- tcp --
+        dict.set_item(intern!(py, "tcp_keepalive"), self.tcp_keepalive)?;
+        dict.set_item(
+            intern!(py, "tcp_keepalive_interval"),
+            self.tcp_keepalive_interval,
+        )?;
+        dict.set_item(
+            intern!(py, "tcp_keepalive_retries"),
+            self.tcp_keepalive_retries,
+        )?;
+        dict.set_item(intern!(py, "tcp_nodelay"), self.tcp_nodelay)?;
+        // -- tls --
+        dict.set_item(intern!(py, "tls_min_version"), self.tls_min_version)?;
+        dict.set_item(intern!(py, "tls_max_version"), self.tls_max_version)?;
+        dict.set_item(intern!(py, "tls_info"), self.tls_info)?;
+        dict.set_item(intern!(py, "tls_sni"), self.tls_sni)?;
+        dict.set_item(
+            intern!(py, "danger_accept_invalid_certs"),
+            self.danger_accept_invalid_certs,
+        )?;
+        dict.set_item(
+            intern!(py, "danger_accept_invalid_hostnames"),
+            self.danger_accept_invalid_hostnames,
+        )?;
         Ok(dict)
     }
 }
@@ -662,7 +991,37 @@ impl ClientConfig {
             .deflate(self.deflate)
             .zstd(self.zstd)
             .cookie_store(self.cookies)
-            .hickory_dns(self.hickory_dns);
+            .hickory_dns(self.hickory_dns)
+            .referer(self.referer)
+            .redirect(
+                self.redirect
+                    .map_or_else(reqwest::redirect::Policy::none, |max| {
+                        reqwest::redirect::Policy::limited(max)
+                    }),
+            )
+            .https_only(self.https_only)
+            .http1_allow_obsolete_multiline_headers_in_responses(
+                self.http1_allow_obsolete_multiline_headers_in_responses,
+            )
+            .http1_allow_spaces_after_header_name_in_responses(
+                self.http1_allow_spaces_after_header_name_in_responses,
+            )
+            .http1_ignore_invalid_headers_in_responses(
+                self.http1_ignore_invalid_headers_in_responses,
+            )
+            .pool_idle_timeout(self.pool_idle_timeout.map(|d| d.0))
+            .pool_max_idle_per_host(self.pool_max_idle_per_host)
+            .tcp_keepalive(self.tcp_keepalive.map(|d| d.0))
+            .tcp_keepalive_interval(self.tcp_keepalive_interval.map(|d| d.0))
+            .tcp_keepalive_retries(self.tcp_keepalive_retries)
+            .tcp_nodelay(self.tcp_nodelay)
+            .tls_sni(self.tls_sni)
+            .tls_info(self.tls_info)
+            .danger_accept_invalid_certs(self.danger_accept_invalid_certs)
+            .danger_accept_invalid_hostnames(self.danger_accept_invalid_hostnames);
+        if self.http1_only {
+            client_builder = client_builder.http1_only();
+        }
         if let Some(user_agent) = &self.user_agent {
             client_builder = client_builder.user_agent(user_agent.clone());
         }
@@ -680,6 +1039,52 @@ impl ClientConfig {
         }
         if self.http1_only {
             client_builder = client_builder.http1_only();
+        }
+
+        // http1
+        if self.http1_title_case_headers {
+            client_builder = client_builder.http1_title_case_headers();
+        }
+
+        // http2
+        if self.http2_prior_knowledge {
+            client_builder = client_builder.http2_prior_knowledge();
+        }
+        if let Some(http2_initial_stream_window_size) = self.http2_initial_stream_window_size {
+            client_builder =
+                client_builder.http2_initial_stream_window_size(http2_initial_stream_window_size);
+        }
+        if let Some(http2_initial_connection_window_size) =
+            self.http2_initial_connection_window_size
+        {
+            client_builder = client_builder
+                .http2_initial_connection_window_size(http2_initial_connection_window_size);
+        }
+        if self.http2_adaptive_window {
+            client_builder = client_builder.http2_adaptive_window(true);
+        }
+        if let Some(http2_max_frame_size) = self.http2_max_frame_size {
+            client_builder = client_builder.http2_max_frame_size(http2_max_frame_size);
+        }
+        if let Some(http2_max_header_list_size) = self.http2_max_header_list_size {
+            client_builder = client_builder.http2_max_header_list_size(http2_max_header_list_size);
+        }
+        if let Some(http2_keep_alive_interval) = &self.http2_keep_alive_interval {
+            client_builder = client_builder.http2_keep_alive_interval(http2_keep_alive_interval.0);
+        }
+        if let Some(http2_keep_alive_timeout) = &self.http2_keep_alive_timeout {
+            client_builder = client_builder.http2_keep_alive_timeout(http2_keep_alive_timeout.0);
+        }
+        if self.http2_keep_alive_while_idle {
+            client_builder = client_builder.http2_keep_alive_while_idle(true);
+        }
+
+        // tls
+        if let Some(tls_min_version) = &self.tls_min_version {
+            client_builder = client_builder.min_tls_version(tls_min_version.into());
+        }
+        if let Some(tls_max_version) = &self.tls_max_version {
+            client_builder = client_builder.max_tls_version(tls_max_version.into());
         }
         client_builder
     }